Install sqlite 3.44.3
This installs sqlite 3.44.3 as a candidate sqlite source release.
This is not the official release - it is 3.44.2 plus two extra commits
that resolve issues found during Android testing.  However, sqlite.org
has named the release "3.44.3".

This change does not change the sqlite release used in the code.  The
trunk-stable build flag RELEASE_PACKAGE_LIBSQLITE3 must be updated to
3440300.

Tested with a temporary CL that advanced the build flag to use the new
source.  Verified that a phone running the build was reporting
"3.44.3" as the sqlite version.

Bug: 322547687

Change-Id: I38e6c37b54ac20fe264b1760f30efae268fc945b

[YAAP]: actually switch to this entirely in dist rather than a
build_flag candidate. Also fix METADATA